pyproject.toml
[tool.poetry]
name = "telegramtagallbot"
version = "0.1.0"
description = ""
authors = ["DmitriyKruchkov <kruchkovdmitriy48@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = {extras = ["ext"], version = "^21.0.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
